v0.0011

+++ main.cpp
@@ -24,12 +24,12 @@ int main() {
                     cout<<"Masukkan NIK : ";
                     cin>>dataCust.NIK;
                     cout<<"Masukkan umur : ";
-                    dataCust.age = validateIntInput(dataCust.age);
+                    cin>>dataCust.age;
                     cout<<"Masukkan jenis kelamin : ";
                     cin>>dataCust.gender;
-                    custP = getCustomer(Customers, dataCust);                    
-                    if(custP != NULL){ // or if(getCustomer(Customers, dataCust) != NULL){...}
-                        cout<<"customer/NIK sudah terdaftar\n";
+                    bool findNIK = getNIK(Customers, dataCust);
+                    if(findNIK == true){
+                        cout<<"NIK sudah terdaftar\n";
                     }else{
                         custP = createElementCustomer(Customers, dataCust);
                         insertLastCustomer(Customers, custP);
@@ -50,8 +50,9 @@ int main() {
                 if(custP != NULL){
                     showCustomerData(Customers, dataCust);
                 }else{
-                    cout<<"Customer tidak terdaftar\n";
+                    cout<<"Masukkan Nama/NIK tidak terdaftar\n";
                 }
+                
                 break;
             }
             case 3:{ // Menghapus customer tertentu [done] [clean]
@@ -73,7 +74,7 @@ int main() {
                 cin>>dataCust.NIK;
                 custP = getCustomer(Customers, dataCust);
                 if(custP == NULL){ // or if(getCustomer(Customers, dataCust) != NULL){...}
-                    cout<<"Customer tidak ditemukan\n";
+                    cout<<"Data customer tidak ditemukan\n";
                 }else{
                     cout<<"Customer ditemukan\n\n";
                     cout<<"Nama\t\t: "<<data(custP).name<<"\n";
@@ -94,20 +95,20 @@ int main() {
                 cin>>dataCust.NIK;
                 custP = getCustomer(Customers, dataCust);
                 if(custP != NULL){ // or if(getCustomer(Customers, dataCust) != NULL){...}
-                    cout<<"Customer berhasil ditemukan\n";
+                    cout<<"Nama customer berhasil ditemukan\n";
                     cout<<"Masukkan nama kredit : ";
                     cin>>dataCred.creditName;
                     cout<<"Masukan nomor rekening : ";
                     cin>>dataCred.creditID;
                     adrCredit credP = getCredit(Credits, dataCred);
                     if(credP != NULL){
-                        cout<<"Kartu kredit sudah ada dalam list\n";
+                        cout<<"Data kredit sudah ada dalam list\n";
                     }else{
                         addCreditToCustomer(Credits, Customers, dataCred, dataCust);
                         cout<<"Kartu kredit berhasil ditambahkan\n";
                     }
                 }else{
-                    cout<<"Customer tidak ditemukan\n";
+                    cout<<"Nama customer tidak ditemukan\n";
                 }
                 cout<<"\n";
             break;
@@ -119,7 +120,12 @@ int main() {
                 cin>>dataCust.name;
                 cout<<"Masukkan NIK customer : ";
                 cin>>dataCust.NIK;
-                deleteAllCreditInCustomer(Credits, Customers, dataCust);
+                custP = getCustomer(Customers, dataCust);
+                if(custP != NULL){
+                    deleteAllCreditInCustomer(Credits, Customers, dataCust);
+                }else{
+                    cout<<"Customer tidak terdaftar\n";
+                }
                 break;  
             }
             case 7: { // Menampilkan seluruh kartu kredit dari customer X [done] [clean]
@@ -142,7 +148,7 @@ int main() {
                 cin>>dataCust.NIK;
                 custP = getCustomer(Customers, dataCust);
                 if(custP != NULL){ // or if(getCustomer(Customers, dataCust) != NULL){...}
-                    cout<<"Customer berhasil ditemukan\n";
+                    cout<<"Nama customer berhasil ditemukan\n";
                     cout<<"Masukkan nama kredit : ";
                     cin>>dataCred.creditName;
                     cout<<"Masukkan nomor rekening : ";
@@ -157,7 +163,7 @@ int main() {
                         cout<<"Kartu kredit tidak ditemukan";
                     }
                 }else{
-                    cout<<"Customer tidak ditemukan\n";
+                    cout<<"Nama customer tidak ditemukan\n";
                 }
                 cout<<"\n";
                 break;
@@ -172,7 +178,7 @@ int main() {
                 cin>>dataCust.NIK;
                 custP = getCustomer(Customers, dataCust);
                 if(custP != NULL){ // or if(getCustomer(Customers, dataCust) != NULL){...}
-                    cout<<"Customer berhasil ditemukan\n";
+                    cout<<"Nama customer berhasil ditemukan\n";
                     cout<<"Masukkan nama kredit : ";
                     cin>>dataCred.creditName;
                     cout<<"Masukan nomor rekening : ";
@@ -180,7 +186,7 @@ int main() {
                     connectCreditToCustomer(Credits, Customers, dataCred, dataCust);
                     cout<<"Berhasil menghubungkan kartu kredit dengan customer\n";
                 }else{
-                    cout<<"Customer tidak ditemukan\n";
+                    cout<<"Nama customer tidak ditemukan\n";
                 }
                 cout<<"\n";
                 break;
@@ -195,7 +201,7 @@ int main() {
                 cin>>dataCust.NIK;
                 custP = getCustomer(Customers, dataCust);
                 if(custP != NULL){ // // or if(getCustomer(Customers, dataCust) != NULL){...}
-                    cout<<"Customer berhasil ditemukan\n";
+                    cout<<"Nama customer berhasil ditemukan\n";
                     if(data(custP).totalCredits > 0){
                         showCustomerCredit(Credits, Customers, dataCust);
                         cout<<"Masukkan nama kredit : ";
@@ -213,7 +219,7 @@ int main() {
                         cout<<"Customer tidak memiliki kartu kredit\n";
                     }
                 }else{
-                    cout<<"Customer tidak ditemukan\n";
+                    cout<<"Nama customer tidak ditemukan\n";
                 }
                 cout<<"\n";
                 break;
@@ -233,7 +239,7 @@ int main() {
                 cin>>dataCred.creditID;
                 credP = getCredit(Credits, dataCred);
                 if(credP == NULL){ // or if(getCredit(Credits, dataCred) != NULL){...}
-                        cout<<"Kredit tidak ditemukan/tidak terdaftar\n";
+                        cout<<"Data kredit tidak ditemukan/tidak terdaftar\n";
                 }else{
                     custP = getCustomerFromCredit(Credits, Customers, dataCred);
                     if(custP == NULL){
